fixup! CHE-6255: Apply 3-state checkbox in Git changes list widget
