+++ plugins/plugin-git/che-plugin-git-ext-git/src/main/java/org/eclipse/che/ide/ext/git/client/compare/selectablechangespanel/SelectableChangesPanelView.java
@@ -25,6 +25,9 @@ public interface SelectableChangesPanelView extends ChangesPanelView {
   /** Needs for delegating actions into {@link SelectableChangesPanelPresenter}. */
   interface ActionDelegate {
 
+    /** Refresh all nodes in the panel. */
+    void refreshNodes();
+
     /** Is called when item check-box changed selection state. */
     void onFileNodeCheckBoxValueChanged(Path path, boolean newCheckBoxValue);
