+++ plugins/plugin-git/che-plugin-git-ext-git/src/main/java/org/eclipse/che/ide/ext/git/client/compare/selectablechangespanel/SelectionCallBack.java
@@ -21,6 +21,7 @@ public interface SelectionCallBack {
 
   /**
    * Is called when item check-box changed selection state.
+   *
    * @param path path of the selected item
    * @param isChecked checkbox selection state
    */
